proxy: v2.183.0 (#8869)
This release updates the proxy to configure inbound route
configurations--including authorizations and filters--via discovery from
the control plane.

If the controller does not yet configure routes, the proxy uses a
default route. Route authorizations always extend the server's
authorizations, so routes need not be defined to admit traffic. However,
if a server have at least one route, *only* traffic that matches a route
will be permitted, otherwise requests fail with a 404.

If the proxy receives a router filter of an unknown type (i.e., because
the controller is running a later version of the API that includes new
types), then the proxy will FAIL all requests on that route with an
internal server error. It's considered safer to fail hard in this case,
rather than to silently ignore a configured policy that could
potentially be security-sensitive.

---

* build(deps): bump smallvec from 1.8.1 to 1.9.0 (linkerd/linkerd2-proxy#1795)
* build(deps): bump serde_json from 1.0.81 to 1.0.82 (linkerd/linkerd2-proxy#1797)
* build(deps): bump either from 1.6.1 to 1.7.0 (linkerd/linkerd2-proxy#1798)
* Add an HTTP route redirect filter (linkerd/linkerd2-proxy#1799)
* Add HTTP and gRPC route failure-injection filters (linkerd/linkerd2-proxy#1805)
* build(deps): bump tracing-subscriber from 0.3.11 to 0.3.14 (linkerd/linkerd2-proxy#1806)
* build(deps): bump pin-project from 1.0.10 to 1.0.11 (linkerd/linkerd2-proxy#1807)
* build(deps): bump unicode-normalization from 0.1.20 to 0.1.21 (linkerd/linkerd2-proxy#1808)
* build(deps): bump fixedbitset from 0.4.1 to 0.4.2 (linkerd/linkerd2-proxy#1809)
* build(deps): bump serde from 1.0.137 to 1.0.138 (linkerd/linkerd2-proxy#1810)
* build(deps): bump regex from 1.5.6 to 1.6.0
* build(deps): bump once_cell from 1.12.0 to 1.13.0
* build(deps): bump hyper from 0.14.19 to 0.14.20 (linkerd/linkerd2-proxy#1816)
* add shutdown grace period timeout for graceful shutdowns (linkerd/linkerd2-proxy#1815)
* dev: Update dev tooling (linkerd/linkerd2-proxy#1820)
* build(deps): bump serde from 1.0.138 to 1.0.139 (linkerd/linkerd2-proxy#1819)
* build(deps): bump hashbrown from 0.12.1 to 0.12.2 (linkerd/linkerd2-proxy#1818)
* build(deps): bump prettyplease from 0.1.15 to 0.1.16 (linkerd/linkerd2-proxy#1821)
* Revert "build(deps): bump tracing-subscriber from 0.3.11 to 0.3.14 (linkerd/linkerd2-proxy#1806)" (#1822)
* Configure inbound HTTP routes via gRPC (linkerd/linkerd2-proxy#1814)

Signed-off-by: Oliver Gould <ver@buoyant.io>