Update value to target the mode
Instead of using a boolean to dictate whether nft-mode should be used,
we now support two different modes: nft and legacy. By default, iptables
will be installed in nft mode. Furthermore, if an incorrect value is
specified, the mode will still be nft.

To back out of the change and use iptables-legacy, operators can specify
'proxyInit.iptablesMode=legacy' when installing (or upgrading) Linkerd.

Signed-off-by: Matei David <matei@buoyant.io>