dev: Pre-load hokay images instead of nginx
Our dev & CI test helpers preload containers onto clusters so the images
need not be pulled at runtime (which, anecdotally, is much slower). When
we changed the policy tests to use hokay instead of nginx, the helpers
were not updated to preload the correct image.

This change also updates the `justfile` to make it easier to build test
clusters with additional configurations.

Signed-off-by: Oliver Gould <ver@buoyant.io>