ci: Only run `viz` test suite when viz code changes
The `viz` test suite is particularly unreliable. A high percentage of
test runs need to be retried due to spurious failures.

Until we can make these tests more robust, this change isolates the
`viz` test suite to only run when the `viz` directory changes or when
the proxy version is updated.

Signed-off-by: Oliver Gould <ver@buoyant.io>