policy: validate that `HTTPRoute` paths are absolute  (#8962)
The proxy won't handle httproute paths (in URI rewrites or matches) when
paths are relative. The policy admission controller and indexer should
catch this case and fail to handle routes that deal in paths that do not
start in `/`.

This branch adds validation to the admission controller and indexer to
ensure that all paths in an `httproute` rule are absolute.