Change `nodeAffinity` type in charts pkg (#8992)
Linkerd supports nodeAffinity for its control plane components. In code,
nodeAffinity is expressed as a map with string keys and string values.
In practice, the types do not appear to be correct, however.

The proxy-injector will, for example, read the install values from a
ConfigMap and unmarshal them into a Values struct. Unmarshalling fails
due to a type error: error unmarshaling JSON: while decoding JSON: json:
cannot unmarshal object into Go struct field Values.nodeAffinity of type
string. While the control plane will be deployed with the correct
affinity values, any subsequent injections will fail as a result of the
incorrect types.

To fix the issue, the type has been changed from map[string]string, to
map[string]interface{}. While the keys are preserved as strings, values
should be arbitrary in this case (since we do not have a concrete
representation of the affinity object).

Signed-off-by: Matei David <matei@buoyant.io>