devcontainer: install Google protobuf include files
This commit updates the devcontainer Dockerfile to also install the
Google protobuf definitions for well-known types to
`/usr/local/include`, and set the `PROTOC_INCLUDE` environment variable
to point at these files in the devcontainer. This *should* fix [this
build failure][1] when compiling OpenCensus protobufs in the proxy repo.

I believe this will also allow us to un-vendor the Google protobuf
definitions in other places (e.g. `linkerd2-proxy-api`), as long as the
builds also happen in the devcontainer Docker image.

If this *isn't* the most correct way to resolve this, I'm happy to
consider other options (e.g., vendoring the protos in the proxy repo as
well), but it felt like the "most proper" solution.

[1]: https://github.com/linkerd/linkerd2-proxy/runs/7677474405?check_suite_focus=true