justfile: redirect `_k3d-dns-ready` to stderr
Currently, the `just linkerd` recipe cannot be used to output YAML
that's piped into another command. This is because the recipe depends on
the `_k3d-dns-ready` recipe, which runs a `kubectl wait` command to
waite for kubedns to be ready. This command outputs a line linkerd
```
pod/coredns-b96499967-p9ct5 condition met
```
to stdout.

When the `just linkerd` command is part of a pipeline, such as
```shell
$ just linkerd install | just k apply -f -
```
the line output by `kubectl wait` is prepended to the Linkerd CLI's
output. Thus, this `kubectl apply` pipeline will fail, since that line
is not valid YAML.

This PR fixes this by redirecting the `kubectl wait` output to stderr,
so that it's not preprended to the Linkerd CLI command's stdout.