dev: Make Go/Rust/Node directories writeable
`cargo`, `go`, and `npm` may try to write to directories that are part
of the devcontainer's base install. This change ensures that these
directories are writeable by all users in the devcontainer.

This change also updates golangci-lint to latest and adds a
`ReadHeaderTimeout` setting to HTTP servers to prevent slowloris
attacks.

Signed-off-by: Oliver Gould <ver@buoyant.io>