Pass TAG into arm64 integration tests (#9146)
The arm64 integration tests require that the TAG env variable is set properly so that they can invoke the correct tag of the cni-plugin image.  However, this env variable is not being set for the integration tests, resulting in the wrong tag being used.

e.g. see https://github.com/linkerd/linkerd2/runs/7810461484?check_suite_focus=true

We set the TAG variable into the GITHUB_ENV so that it is available to the integration test.

Signed-off-by: Alex Leong <alex@buoyant.io>