Add config `proxyInit.runAsUser` to facilitate 2.11.x->2.12.0 upgrade.
In `2.11.x`, `proxyInit.runAsRoot` was true by default, which caused the
proxy-init's `runAsUser` field to be 0. `proxyInit.runAsRoot` is now
defaulted to false in `2.12.0`, but `runAsUser` still isn't
configurable, and when following the upgrade instructions
[here](https://github.com/linkerd/website/blob/f848bd53a0f5d9f5a0ea795679d3dcc66d1333b1/linkerd.io/content/2.12/tasks/upgrade.md#upgrade-notice-stable-2120),
helm doesn't change `runAsUser` and so it conflicts with the new value
for `runAsRoot=false`, resulting the pods erroring with this message:
```
Error: container's runAsUser breaks non-root policy (pod: "linkerd-identity-bc649c5f9-ckqvg_linkerd(fb3416d2-c723-4664-acf1-80a64a734561)", container: linkerd-init)
```

This PR adds a new default for `runAsUser` to avoid this issue.