justfile: allow passing additional args to `linkerd install`
The justfile has a `linkerd-install` recipe that installs Linkerd in the
test k3d cluster, with the appropriate flags to use locally built Docker
image versions. This is useful when testing changes to Linkerd manually,
as well as when running automated tests. In some cases, though, testing
a change may require passing arguments to the `linkerd install` command,
such as a default inbound policy, but this recipe doesn't currently
provide a way to do this.

This branch updates the Justfile so that the `linkerd-install` recipe
can be provided with an optional set of additional arguments, which are
passed to `linkerd install`.