refactor use of `watch` channels in policy controller index
Currently, the policy-controller index's `PodPortServer` struct stores
both the `Sender` and `Receiver` sides of a `tokio::sync::watch`
channel. This is unnecessary, since `Sender` has a `subscribe` method
that creates a new `Receiver`, and a `send_if_modified` method for
conditionally checking whether the value in the `watch` channel should
be updated.

This branch refactors the `PodPortServer` type to use these `Sender`
methods. This doesn't result in any functional change, but is maybe a
little bit nicer. Not a big deal though.