Fatal checks should not be warnings (#9493)
Linkerd-viz has some healthchecks which are both fatal and warnings.  This can cause healthchecks to terminate early (because of a failed fatal check) but not count as an error which can cause invalid states where e.g. the apiClient is nil (as in https://github.com/linkerd/linkerd2/issues/9416).  Furthermore, a fatal warning is a bit contradictory.

We change these checks to be fatal errors instead.

Signed-off-by: Alex Leong <alex@buoyant.io>