Move pause container declaration to values (#9566)
For our multicluster gateway we use a pause container as the "main"
container in the pod. Currently, the image version and registry are both
hardcoded into the template itself. This makes a bit difficult to write
CI logic and just recipes.

To keep everything simpler (and neater) to use, this change decouples
the registry and image of the pause container from the template, by
having a unified declaration in the multicluster values file.

This should unblock and help with PR#9135 (run mc tests on min and max
k8s versions)

Signed-off-by: Matei David <matei@buoyant.io>