Remove _pause-image variable from justfile
Just does not support lazy evaluation of commands when their output is
stored in a variable. `_pause-load` was used to store the output of a
`yq` command that would fetch the pause image used in the gateway
template for multicluster.

When running workloads in CI that do not run in the devcontainer, yq is
missing as a dependency. Since the command is eagerly evaluated, we get
an error (even when the recipe that makes use of the variable is not
invoked). This change removes the variable and hardcodes the command in
the recipe itself.

To work around parsing issues for multi-line constructs, we also escape
newlines with slashes ('\').

Signed-off-by: Matei David <matei@buoyant.io>