helm: use `proxyInit`'s user for the `noop` init container
Currently, the `noop` init container created by the Linkerd CNI plugin
causes issues when a workload with
```yaml
securityContext:
  runAsNonRoot: true
```
is injected, since it will add a container that runs as root to that
workload.

This branch resolves this issue by changing the Helm template for the
noop init container to use the same user as the `proxyInit` init
container. I've tested this by injecting a deployment with the above
`securityContext` configuration and verifying that the `noop` init
container is now allowed to run.

This PR is against the `release/stable-2.12` branch, as the `noop` init
container has been removed on the edge branch (as it was replaced with
the CNI validator init container).

Fixes #9671