Tidy go.mod file
The last few commits are compromised of Go dependency updates.
Everything seems to work well, expect for the CLI, which can no longer
be built. This holds for both CI and local builds. The error messages
suggest that the go.mod file should be tidied.

For example, trying to do `/bin/build-cli-bin` fails with:

```
```

After running `go mod tidy`, the CLI seems to build. This should fix
in-flight dependency bumps and unblock the CI for other pull requests.

Signed-off-by: Matei David <matei@buoyant.io>