Use self-hosted runner for ARM64 integration tests
This moves the "ARM64 integration tests" job in `relase.yaml` to a
separate custom action that is run in an ARM self-hosted runner tagged
with `[self-hosted, Linux, ARM64]`, tied to the linkerd github org.

We no longer use a local (linux/x86_64) linkerd CLI that connects to an
existing k3s instance in the host. Instead, we run the CLI ARM64 binary
in the host itself, after creating the cluster with k3d (which gets
always torn down at the end of the tests regardless of their success).

This action can be easily triggered and tested through a separate
workflow like this one:

```yaml
name: eraseme - just for testing the arm64-integration action

on: push

jobs:
  arm64_integration_tests:
    runs-on: [self-hosted, Linux, ARM64]
    timeout-minutes: 30
    steps:
      - uses: actions/checkout@93ea575cb5d8a053eaa0ac8fa3b40d7e05a33cc8
      - uses: ./.github/actions/arm64-integration
        with:
          tag: edge-22.11.1
          docker-registry: ghcr.io/linkerd
          k3d-version: v5.4.4

```

The results for that workflow test are here:
https://github.com/linkerd/linkerd2/actions/runs/3464532975/jobs/5786211951
(run took 2m20s to complete)

Please check the "ARM CI host at Equinix Metal" doc in Notion for the
host setup.

## Other Changes

- The cni test was removed.
- Replaced `"$bindir"/docker` with just `docker` in `bin/image-load` as
  we do elsewhere.
- Properly detect k3d arch in `bin/k3d`