Use self-hosted runner for ARM64 integration tests (#9830)
* Use self-hosted runner for ARM64 integration tests

This refactors the "ARM64 integration tests" job in `relase.yaml` to
use an ARM self-hosted runner tagged with `[self-hosted, Linux, ARM64]`,
tied to the linkerd github org.

We no longer use a local (linux/x86_64) linkerd CLI that connects to an
existing k3s instance in the host. Instead, we run the CLI ARM64 binary
in the host itself, after creating the cluster with k3d (which gets
always torn down at the end of the tests regardless of their success).

Please check the "ARM CI host at Equinix Metal" doc in Notion for the
host setup.

## Other Changes

- The cni test was removed.
- Replaced `"$bindir"/docker` with just `docker` in `bin/image-load` as
  we do elsewhere.
- Properly detect k3d arch in `bin/k3d`