Resuscitate `helm-upgrade` integration test
Based off of the `alpeb/ext-cli-version-override` (depends on #9855).

This had been commented-out when we jumped to the new 2.12 helm charts.
It can be used again to test upgrades from 2.12.x.

- Some of the logic in `test/integration/install/install_test.go` still
  hadn't considered the need to upgrade both the `linkerd-crds` and
  `linkerd-control-plane` charts, so that got fixed.
- Removed references to the now-deprecated `linkerd2` chart.
- Improved the `helm_cleanup()` function by uninstalling the charts in
  reverse order (extensions first, core last). We delete the namespaces
  afterwards because helm sometimes doesn't remove them, and so we
  shouldn't fail if we attempt to delete one that is already gone. Also
  removed unneeded `kubectl wait`s because `kubect delete ns` should be
  blocking.

Disclaimer: This is still not an ideal test setup, but it was a minor
effort getting the existing code fixed before migrating to an approach
less dependent on bash.